refactor: rename Breadcrumbs component
